Conversion to a different nextToken implementation (#6)
* Start of conversion to a different nextToken implementation

This looks like more changes than there actually are, since the addition of a while loop changed the indentation level within nextToken.

The basic ideas:
- Removal of Token.EndOfFile, instead nextToken returning null is *always* EOF
  + nextToken will loop until it has either an error or a token ready for consumption instead of returning intermediate null tokens
  + Whenever we know a token or error exists, we return it from the nextToken function as soon as we're done setting up the tokenizer state for the next character/token
- Addition of Tokenizer.errorQueue that can be used to return multiple errors at once (https://github.com/watzon/zhtml/issues/3)

Benefits:
- Allows for more straightforward handling on the caller's side of nextToken, makes it more similar to other next functions in Zig that return null once they are finished iterating
- Allows for implementing something like "When an end tag token is emitted with attributes, that is an end-tag-with-attributes parse error." from the spec without a lot of special-casing, since we can just append the errors in emitToken and know that we're going to immediately return the errors afterwards

Drawbacks:
- Potentially leads to a more complicated to follow nextToken implementation since it's a loop rather than a single switch per call

* Don't compare EndTag attributes in expectEqualTokens

EndTag attributes are ultimately ignored; we only store them to be able to emit EndTagWithAttributes errors

* Integrate character references with popQueuedErrorOrToken paradigm

* Move eof check inside the loop

* Make netChar/peekChar/etc handle EOF and add tests

* Reintroduce Token.EndOfFile, make nextToken never return null

Didn't realize the spec actually says 'emit an end-of-file token', so having an actual EndOfFile token makes more sense to me than returning null as an EndOfFile surrogate. Having it be an actual token also allows for adding EndOfFile to the backlog if that is every necessary somewhere.

* Example of handling EOF in a state

All the other states will need similar handling (for every call of nextChar/peekChar/peekN). This is just an example

* Update the rest of the switch to handle EOF

This just gets it to compile again, things are still quite broken

* Rework next/peek/current again to actually be right this time

My last attempt at this was not right at all. It didn't handle the beginning/end anywhere close to properly

* Add missing '>' case to ScriptDataEscapedDashDash state

* A few fixes

* Fix peekN when reconsuming and fix test

* Use hasQueuedErrorOrToken instead of bool return

* Remove useless std.math.min call

* Fix peekChar for second-to-last character